pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "novel-generator"
version = "0.1.0"
description = "Self-hosted Ollama-first long-form fiction generation studio."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "RhythmicCarnage" }
]
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "python-docx>=1.1.2,<2.0.0",
  "python-multipart>=0.0.12,<1.0.0",
  "pydantic-settings>=2.6.1,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3,<9.0.0",
]

[project.scripts]
novel-generator-web = "novel_generator.main:run"
novel-generator-worker = "novel_generator.worker:run"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
novel_generator = ["templates/*.html", "static/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
